crawler/requirements.txt
pymysql>=1.1.0
pytest>=7.4.0
pytest-dotenv>=0.5.2
python-dotenv>=1.0.0
Appium-Python-Client>=3.1.0
paddleocr>=2.7.0
Pillow>=10.0.0
